+++ go/internal/converters.go
@@ -575,3 +575,7 @@ func ConverterAndTypeChecker(
 	// TODO maybe still return the data?
 	return nil, fmt.Errorf("unexpected return type from Glide: got %v, expected %v", reflect.TypeOf(data), expectedType)
 }
+
+func ConvertKeyValuesArrayOrNil(data any) ([]models.KeyValues, error) {
+	return keyValuesConverter{canBeNil: true}.convert(data)
+}